refactor: client builder and api details setter
