+++ src/services/customerCreator.tsx
@@ -1,11 +1,10 @@
-import { apiRoot } from './client';
+import { apiRoot } from './clientBuilder';
 import { CustomerDraft } from '../../utils/types';
 
 export function createCustomer(customerDetails: CustomerDraft) {
   return apiRoot
     .customers()
     .post({
-      // The CustomerDraft is the object within the body
       body: customerDetails,
     })
     .execute();
